Handle zero-length vectors properly
The Prometheus HTTP handlers expect all protobuf `MetricFamily` structs
to actually contain some metrics, so we need to filter out zero-length
vectors when gathering our metrics.

Fixes #17.
