+++ core.go
@@ -56,7 +56,10 @@ func newCore() *core {
 		c.RLock()
 		protos := make([]*promproto.MetricFamily, 0, len(c.metrics))
 		for _, m := range c.metrics {
-			protos = append(protos, m.proto())
+			p := m.proto()
+			if p != nil && len(p.Metric) > 0 {
+				protos = append(protos, p)
+			}
 		}
 		c.RUnlock()
 		return protos, nil
